Progress, at a price. To be improved.
Using one definition requires a verification of
(conv a c) that can only be proved by transitivity
via (conv a b) and (conv b c). We introduce a
(conv_trans a b c) that asks both to hold and we use
an hyp to ask the conv_trans when the definition is used
and assume (conv a c) during its verification.

In the future we need to prove it once and for all, not
every time the hypothesis is instantiated.