Primitive projections p1,p2 for setSigma
I have changed the calculus introducing primitive
projections p1 and p2 to greatly reduce the size of
generated terms. The code is also simplified.

For the time being I keep both the projections and the
eliminator and I have not defined the first to be
reducible to the latter.

An experiment on defining them as library objects failed
because of the way the extensional level look-up for proofs
for (propEq _ A B).