Major change: reduction no longer perform typing.
As in Coq, Matita, etc. we assume terms to be convertible/reduced
to have already been typed. (And, in case of conversion, we assume
their types to have been already checked to be convertible).

This gives a major speed-up (from quadratic to linear factor).
Moreover the old version defeated the existence of a library where
terms were type-checked once and for all.

The fact that the two versions of the theories are equivalent and
that the invariant is always respected should be done formally.